requirements.txt
openai==0.27.0
python-dotenv
pandas==2.2.2
Django==5.0.6
djangorestframework==3.15.2